Read serial number form files instead of statusamodule.
Reading the serial number from the status module makes all modules
dependign on the status module. Reading the serialnumber file instead
removes this dependency.

The lincesystem class got an additional
parameter inside the constructor with ther serialnumber filepath.
The licenseSystem is not a evetnsystem any longer because it does not
need to receive information from the status module.

Using this debugging is only possible with a master license, when
zera-modulemanger is not running on the target.

Signed-off-by: bhamacher <b.hamacher@zera.de>